EventSystem: added an event system
-This will allow a lot of decoupling and removing hardcoded references to other GameObjects.
Want to know when the HP of the player changed? Add a FloatListener and listen on that particular event. Now the player doesn't have to care who needs to know about his HP, he simply raises an event when it's changed. Hopefully this makes it easy to program communication between systems (player/spell casting/enemies(etc...)